# openapi/openapi.yaml
openapi: 3.0.3
info:
    title: AdGuard Home
    description: AdGuard Home REST API. Admin web interface is built on top of this REST API.
    version: "0.104"
    contact:
        name: "AdGuard Home"
        url: "https://github.com/AdguardTeam/AdGuardHome"
servers:
    - url: /control
...
paths:
...
    /filtering/config:
        post:
            tags:
                - filtering
            operationId: filteringConfig
            summary: Set filtering parameters
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/FilterConfig"
                required: true
            responses:
                "200":
                    description: OK
...
components:
...
    schemas:
...
        FilterConfig:
            type: object
            description: Filtering settings
            properties:
                enabled:
                    type: boolean
                interval:
                    type: integer
...